+++ experimental/dds/tree2/src/test/feature-libraries/default-field-kinds/optionalChangeRebaser.spec.ts
@@ -41,7 +41,7 @@ import {
 	generatePossibleSequenceOfEdits,
 	ChildStateGenerator,
 } from "../../exhaustiveRebaserUtils";
-import { runExhaustiveComposeRebaseSuite } from "../rebaserAxiomaticTests";
+import { runExhaustiveComposeRebaseSuite } from "../../rebaserAxiomaticTests";
 
 type RevisionTagMinter = () => RevisionTag;
